Clean up GitTrackObject controller
